fix: twitter_simple shortcode, paginator issue since hugo `v0.123.0` (#170)
* fix: twitter_simple shortcode, closes #171
* bump HUGO_VERSION to 0.123.7 in netlify.toml
* fix paginator issue since hugo v0.123.0

Pages of kind "page" don't allow the creation of a paginator (.Paginate
or .Paginator) in the newest version of hugo.
Pages of kind "section" (e.g. /posts) still allow the creation. This fix
adds a check for the kind of page and excludes pages with kind "page"
from creating a paginator.


---------

Co-authored-by: Sidharth R <122173059+hugo-sid@users.noreply.github.com>